Rename stream.hpp to cuda_stream_view.hpp
