+++ include/rmm/mr/device/hybrid_memory_resource.hpp
@@ -16,8 +16,8 @@
 #pragma once
 
 #include <rmm/detail/aligned.hpp>
+#include <rmm/detail/cuda_stream_view.hpp>
 #include <rmm/detail/error.hpp>
-#include <rmm/detail/stream.hpp>
 #include <rmm/mr/device/device_memory_resource.hpp>
 
 #include <utility>
