+++ include/rmm/mr/device/managed_memory_resource.hpp
@@ -17,8 +17,8 @@
 
 #include "device_memory_resource.hpp"
 
+#include <rmm/detail/cuda_stream_view.hpp>
 #include <rmm/detail/error.hpp>
-#include <rmm/detail/stream.hpp>
 
 #include <cuda_runtime_api.h>
