+++ include/rmm/mr/device/pool_memory_resource.hpp
@@ -15,8 +15,8 @@
  */
 #pragma once
 
+#include <rmm/detail/cuda_stream_view.hpp>
 #include <rmm/detail/error.hpp>
-#include <rmm/detail/stream.hpp>
 #include <rmm/mr/device/detail/free_list.hpp>
 #include <rmm/mr/device/device_memory_resource.hpp>
