+++ include/rmm/mr/device/thread_safe_resource_adaptor.hpp
@@ -15,8 +15,8 @@
  */
 #pragma once
 
+#include <rmm/detail/cuda_stream_view.hpp>
 #include <rmm/detail/error.hpp>
-#include <rmm/detail/stream.hpp>
 #include <rmm/mr/device/device_memory_resource.hpp>
 
 #include <mutex>
